requirements.txt
apns2==0.3.0
asn1crypto==0.23.0
cffi==1.11.0
cryptography==2.0.3
Django==1.11.5
django-filter==1.0.4
django-push-notifications==1.5.0
djangorestframework==3.6.4
enum34==1.1.6
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
ipaddress==1.0.18
Markdown==2.6.9
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
pytz==2017.2
six==1.11.0
